ANALYSIS AND DISCUSION:  QUESTIONNIRE SURVEY
First section details the findings related to the reliability and validity test trough pilot survey that justifies to conduct the survey on the sample as per sampling distribution. Second section identifies the descriptive statistics that validates the sampling distribution in accordance with the research methodology. Third section identifies the key variables that significantly influence customer satisfaction and loyalty in Biman Bangladesh Airlines Limited. Fourth section recognizes the impact of improved customer service on customer loyalty towards Biman Bangladesh Airlines. Fifth section provide suggestion of the respondents about the area of improvement to retain passengers in Biman Bangladesh Airlines Limited. Finally, sixth section summarizes the overall findings and draws conclusions in the light of the findings of the survey.